pyproject.toml
[project]
name = "fastvlm-service"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pillow>=10.0",
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.28",
    "huggingface-hub>=0.22",
]
